User auth, store recipe/user data
Successfully stores some essential user data in a recipe document (which will be displayed in the future). Also successfully stores a string, representing a recipe ID for each recipe that the specific user has created, in an array stored in the user's document. Same for questions created by the specific user.